Move text rendering to Common.
