+++ Common/Render/Text/draw_text_qt.h
@@ -3,21 +3,16 @@
 #include "ppsspp_config.h"
 
 #include <map>
-#include "gfx_es2/draw_text.h"
+#include "Common/Render/Text/draw_text.h"
 
-#if defined(_WIN32) && !defined(USING_QT_UI) && !PPSSPP_PLATFORM(UWP)
+#if defined(USING_QT_UI)
 
-#define WIN32_LEAN_AND_MEAN
-#include <Windows.h>
+#include <QtGui/QFont>
 
-struct TextDrawerContext;
-// Internal struct but all details in .cpp file (pimpl to avoid pulling in excessive headers here)
-class TextDrawerFontContext;
-
-class TextDrawerWin32 : public TextDrawer {
+class TextDrawerQt : public TextDrawer {
 public:
-	TextDrawerWin32(Draw::DrawContext *draw);
-	~TextDrawerWin32();
+	TextDrawerQt(Draw::DrawContext *draw);
+	~TextDrawerQt();
 
 	uint32_t SetFont(const char *fontName, int size, int flags) override;
 	void SetFont(uint32_t fontHandle) override;  // Shortcut once you've set the font once.
@@ -30,12 +25,10 @@ public:
 
 protected:
 	void ClearCache() override;
-	void RecreateFonts();  // On DPI change
-
-	TextDrawerContext *ctx_;
-	std::map<uint32_t, std::unique_ptr<TextDrawerFontContext>> fontMap_;
 
 	uint32_t fontHash_;
+	std::map<uint32_t, QFont *> fontMap_;
+
 	std::map<CacheKey, std::unique_ptr<TextStringEntry>> cache_;
 	std::map<CacheKey, std::unique_ptr<TextMeasureEntry>> sizeCache_;
 };
