+++ UWP/CommonUWP/CommonUWP.vcxproj
@@ -439,6 +439,11 @@
     <ClInclude Include="..\..\Common\Profiler\Profiler.h" />
     <ClInclude Include="..\..\Common\Render\DrawBuffer.h" />
     <ClInclude Include="..\..\Common\Render\TextureAtlas.h" />
+    <ClInclude Include="..\..\Common\Render\Text\draw_text.h" />
+    <ClInclude Include="..\..\Common\Render\Text\draw_text_android.h" />
+    <ClInclude Include="..\..\Common\Render\Text\draw_text_qt.h" />
+    <ClInclude Include="..\..\Common\Render\Text\draw_text_uwp.h" />
+    <ClInclude Include="..\..\Common\Render\Text\draw_text_win.h" />
     <ClInclude Include="..\..\Common\Serialize\Serializer.h" />
     <ClInclude Include="..\..\Common\Serialize\SerializeDeque.h" />
     <ClInclude Include="..\..\Common\Serialize\SerializeFuncs.h" />
@@ -554,6 +559,11 @@
     <ClCompile Include="..\..\Common\Profiler\Profiler.cpp" />
     <ClCompile Include="..\..\Common\Render\DrawBuffer.cpp" />
     <ClCompile Include="..\..\Common\Render\TextureAtlas.cpp" />
+    <ClCompile Include="..\..\Common\Render\Text\draw_text.cpp" />
+    <ClCompile Include="..\..\Common\Render\Text\draw_text_android.cpp" />
+    <ClCompile Include="..\..\Common\Render\Text\draw_text_qt.cpp" />
+    <ClCompile Include="..\..\Common\Render\Text\draw_text_uwp.cpp" />
+    <ClCompile Include="..\..\Common\Render\Text\draw_text_win.cpp" />
     <ClCompile Include="..\..\Common\Serialize\Serializer.cpp" />
     <ClCompile Include="..\..\Common\ColorConv.cpp" />
     <ClCompile Include="..\..\Common\ColorConvNEON.cpp" />
