+++ UWP/NativeUWP/NativeUWP.vcxproj
@@ -379,14 +379,10 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\ext\native\gfx_es2\draw_text.h" />
-    <ClInclude Include="..\..\ext\native\gfx_es2\draw_text_uwp.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\ext\native\gfx_es2\draw_text.cpp" />
-    <ClCompile Include="..\..\ext\native\gfx_es2\draw_text_uwp.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
